asset-generation/pyproject.toml
[project]
name = "asset-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.73.1",
    "pydantic>=2.13.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
]
